[refactor] :recycle: rename utils to helpers folder (#276)
