+++ scripts/subscribe_wallets_to_push_channel.ts
@@ -9,8 +9,8 @@ import * as PushAPI from '@pushprotocol/restapi';
 import { ENV } from '@pushprotocol/restapi/src/lib/constants';
 
 import { IUser } from '../src/models/user';
-import { Logger } from '../src/utils/loggerHelper';
-import { generatePrivateKey } from '../src/utils/keyGeneratorHelper';
+import { Logger } from '../src/helpers/loggerHelper';
+import { generatePrivateKey } from '../src/helpers/keyGeneratorHelper';
 
 dotenv.config();
