+++ scripts/update_users_notif_language.ts
@@ -5,7 +5,7 @@
 import dotenv from 'dotenv';
 import mongoose from 'mongoose';
 
-import { Logger } from '../src/utils/loggerHelper';
+import { Logger } from '../src/helpers/loggerHelper';
 
 dotenv.config();
