+++ src/controllers/nftController.ts
@@ -3,19 +3,18 @@ import mongoose, { ObjectId } from 'mongoose';
 import { FastifyReply, FastifyRequest } from 'fastify';
 
 import { IUser } from '../models/user';
-import { Logger } from '../utils/loggerHelper';
-
+import { Logger } from '../helpers/loggerHelper';
 import { SIGNING_KEY } from '../constants/environment';
-import { getDynamicGas } from '../utils/dynamicGasHelper';
+import { isValidUrl } from '../helpers/validationHelper';
+import { getDynamicGas } from '../helpers/dynamicGasHelper';
 import NFTModel, { INFT, INFTMetadata } from '../models/nft';
 import { getNetworkConfig } from '../services/networkService';
 import { createUserWithWallet } from '../services/userService';
 import { getWalletByPhoneNumber } from '../services/walletService';
 import { sendMintNotification } from '../services/notificationService';
 import { defaultNftImage, networkChainIds } from '../constants/blockchain';
-import { returnErrorResponse, returnSuccessResponse } from '../utils/responseFormatterHelper';
 import { uploadToICP, uploadToIpfs, downloadAndProcessImage } from '../services/uploadService';
-import { isValidUrl } from '../utils/validationHelper';
+import { returnErrorResponse, returnSuccessResponse } from '../helpers/responseFormatterHelper';
 
 export interface NFTInfo {
   description: string;
