+++ src/controllers/withdrawController.ts
@@ -2,7 +2,7 @@ import { ethers } from 'ethers';
 import { FastifyReply, FastifyRequest } from 'fastify';
 
 import { withdrawWalletAllFunds } from '../services/transferService';
-import { returnErrorResponse, returnSuccessResponse } from '../utils/responseFormatterHelper';
+import { returnErrorResponse, returnSuccessResponse } from '../helpers/responseFormatterHelper';
 
 /**
  * Handles the withdrwal all funds
