+++ src/helpers/keyGeneratorHelper.ts
