+++ src/middleware/authMiddleware.ts
@@ -1,8 +1,27 @@
 import { FastifyReply, FastifyRequest } from 'fastify';
 
 import { isPublicRoute } from '../config/publicRoutes';
-import { verifyToken, TokenResponse } from '../config/token';
-import { returnErrorResponse } from '../utils/responseFormatterHelper';
+import { returnErrorResponse } from '../helpers/responseFormatterHelper';
+import { FRONTEND_TOKEN, CHATIZALO_TOKEN } from '../constants/environment';
+
+/**
+ * Represents the possible token types that can be verified
+ */
+type TokenResponse = 'frontend' | 'chatizalo' | null;
+
+/**
+ * Verifies the provided token against known tokens
+ * @param {string} providedToken - The token to verify
+ * @returns {Promise<TokenResponse>} The type of token if verified, or null if not
+ */
+async function verifyToken(providedToken: string): Promise<TokenResponse> {
+  let res: TokenResponse = null;
+
+  if (providedToken === FRONTEND_TOKEN) res = 'frontend';
+  if (providedToken === CHATIZALO_TOKEN) res = 'chatizalo';
+
+  return res;
+}
 
 /**
  * Middleware function to authenticate requests using a Bearer token.
