+++ src/services/predictWalletService.ts
@@ -1,15 +1,14 @@
 import { ethers } from 'ethers';
 import * as crypto from 'crypto';
 
-import { Logger } from '../utils/loggerHelper';
+import { Logger } from '../helpers/loggerHelper';
 import { IBlockchain } from '../models/blockchain';
 import { getNetworkConfig } from './networkService';
 import { SIGNING_KEY } from '../constants/environment';
-
+import { getDynamicGas } from '../helpers/dynamicGasHelper';
 import { getChatterPayWalletFactoryABI } from './bucketService';
-import { generatePrivateKey } from '../utils/keyGeneratorHelper';
-import { ChatterPayWalletFactory, ChatterPayWalletFactory__factory } from '../types/ethers-contracts';
-import { getDynamicGas } from '../utils/dynamicGasHelper';
+import { generatePrivateKey } from '../helpers/keyGeneratorHelper';
+import { ChatterPayWalletFactory__factory } from '../types/ethers-contracts';
 
 export interface PhoneNumberToAddress {
   hashedPrivateKey: string;
@@ -100,4 +99,3 @@ export async function computeProxyAddressFromPhone(phoneNumber: string): Promise
     privateKeyNotHashed: ownerAddress.privateKey
   };
 }
-
