+++ src/services/transferService.ts
@@ -3,8 +3,8 @@ import { ethers } from 'ethers';
 import { IUser } from '../models/user';
 import { getUser } from './userService';
 import { IToken } from '../models/token';
-import { Logger } from '../utils/loggerHelper';
 import Transaction from '../models/transaction';
+import { Logger } from '../helpers/loggerHelper';
 import { IBlockchain } from '../models/blockchain';
 import { getTokenBalances } from './walletService';
 import { setupERC20 } from './contractSetupService';
